A question to confirm deletion of data has been added.

+++ src/command/command.js
@@ -108,7 +108,14 @@ yargs
             if (yargs.argv._.length < 2) {
                 console.log('Api key not found')
             } else if (yargs.argv._.length === 2) {
-                purgeContentObjects(argv.flotiqApiKey, argv.withInternal);
+                const answers = await askQuestions(questionsText.PURGE_QUESTION);
+                const {confirmation} = answers;
+                if (confirmation.toUpperCase() === 'Y') {
+                    await purgeContentObjects(argv.flotiqApiKey, argv.withInternal);
+                } else {
+                    console.log('I\'m finishing, no data has been deleted');
+                    process.exit(1);
+                }
             }
         })
     .help()
@@ -137,11 +144,15 @@ async function checkAllParameters(answer, questions) {
     for (let i = 0; i < questions.length; i++) {
         let paramName = questions[i].name;
         while (!newAnswer[paramName].length) {
-            yargs.showHelp();
-            const param = await inquirer.prompt(questions[i]);
-            newAnswer[paramName] = param[paramName];
-            console.log(newAnswer[paramName]);
-
+            if (!questions[i].defaultAnswer) {
+                yargs.showHelp();
+                const param = await inquirer.prompt(questions[i]);
+                newAnswer[paramName] = param[paramName];
+                console.log(newAnswer[paramName]);
+            } else {
+                newAnswer[paramName] = questions[i].defaultAnswer;
+                console.log(newAnswer[paramName]);
+            }
         }
     }
     return newAnswer;
